app/requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
sqlalchemy==2.0.18
asyncpg==0.25.0
psycopg2==2.9.3
python-dotenv==0.20.0
pydantic==1.10.7
tqdm==4.64.1
